manajer setengah lagi

+++ app/Http/Controllers/Admin/Quiz/QuizController.php
@@ -45,19 +45,23 @@ class QuizController extends Controller
                 ->addIndexColumn()
                 ->addColumn('action', function ($data) {
                     // $btn_action = '<a href="' . route('admin.quiz.show', ['quiz' => $data->id]) . '" class="btn btn-sm btn-info my-1"><i class="fas fa-eye"></i></a>';
+                    $btn_action = '<div align="center">';
+                    if (User::find(Auth::user()->id)->hasAnyRole('admin', 'question-operator')) {
+                        if ($data->type_aspect == 'kecermatan') {
+                            $btn_action .= '<a href="' . route('master.kecermatan.edit', ['quiz' => $data->id]) . '" class="btn btn-sm btn-warning my-1 ml-1"><i class="fas fa-pencil-alt"></i></a>';
+                        } else {
+                            $btn_action .= '<a href="' . route('admin.quiz.edit', ['quiz' => $data->id]) . '" class="btn btn-sm btn-warning my-1 ml-1"><i class="fas fa-pencil-alt"></i></a>';
+                        }
 
-                    if ($data->type_aspect == 'kecermatan') {
-                        $btn_action = '<a href="' . route('master.kecermatan.edit', ['quiz' => $data->id]) . '" class="btn btn-sm btn-warning my-1 ml-1"><i class="fas fa-pencil-alt"></i></a>';
-                    } else {
-                        $btn_action = '<a href="' . route('admin.quiz.edit', ['quiz' => $data->id]) . '" class="btn btn-sm btn-warning my-1 ml-1"><i class="fas fa-pencil-alt"></i></a>';
+                        if ($data->type_aspect != 'kecermatan') {
+                            $btn_action .= '<a href="' . route('admin.quiz.showQuestion', ['quiz' => $data->id]) . '" class="btn btn-sm btn-info my-1 ml-1"><i class="fas fa-search"></i></a>';
+                        }
+
+                        $btn_action .= '<button onclick="destroyRecord(' . $data->id . ')" class="btn btn-sm btn-danger my-1 ml-1"><i class="fas fa-trash"></i></button>';
                     }
-
-                    if ($data->type_aspect != 'kecermatan') {
-                        $btn_action .= '<a href="' . route('admin.quiz.showQuestion', ['quiz' => $data->id]) . '" class="btn btn-sm btn-info my-1 ml-1"><i class="fas fa-search"></i></a>';
-                    }
-
                     $btn_action .= '<a href="' . route('admin.quiz.start', ['quiz' => encrypt($data->id)]) . '" class="btn btn-sm btn-success my-1 ml-1"><i class="fas fa-play"></i></a>';
-                    $btn_action .= '<button onclick="destroyRecord(' . $data->id . ')" class="btn btn-sm btn-danger my-1 ml-1"><i class="fas fa-trash"></i></button>';
+
+                    $btn_action .= '</div>';
                     return $btn_action;
                 })
                 ->only(['name', 'type_aspect', 'action'])
